requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Pillow==10.0.0
tensorflow==2.13.0
numpy==1.24.3
bing-image-downloader==1.1.2
matplotlib==3.7.2
